requirements.txt
bs4
langchain
openai
requests
langchain-openai
langchain-chroma
python-dotenv
unstructured